+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Add version 7 (Quantum Break)
+
 Make sure installed [.Net 5.0](https://dotnet.microsoft.com/download/dotnet/5.0)
 
 Read [here](https://github.com/eprilx/NorthlightFontMaker#usage) for some usage.